CHATROOM/requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.4
django-jalali-date==2.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.2.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
pillow==11.3.0
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
setuptools==80.9.0
sqlparse==0.5.3
uritemplate==4.2.0
